backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-jose
python-keycloak
pydantic